requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-Talisman==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
PyMySQL==1.1.1
WTForms==3.2.1
marshmallow==3.23.2
Gunicorn==23.0.0
python-dotenv==1.0.1
cryptography==42.0.8
